Add parallelism to checking in --upgrade system
The approach here is to do two passes over the deps list. The first converts the
zipper to a clojure map so that we can iterate with standard tools like pmap to
check for new versions. The second loops over the zipper, inserting the
appropriate changes found in the first step.