Dogfood: Automate dependency checking
tl;dr I created a workflow that checks the dependencies of the project
using depot and creates a PR to update the dependencies.

I figured that it would be nice for depot to keep an eye on itself and
check if any of its dependencies get stale.

I don't know what your feelings are towards GitHub actions, but seeing
how the project is on GitHub it made sense to me to create an
automated workflow which would check the dependencies of depot, using
depot.

In the PR I've included the steps for creating the badge that can be
displayed in the README showing if the dependencies are up to date or
not. But before I go and do that, do is this something you think that
the project could use?

If you like this change, maybe I can add one that runs tests also,
that way the code can be checked each time somebody makes a PR or you
merge into master.

What do you think?